Embed match info into index (#312)
* Match information is now embedded (and read from) the search index.
This still has the (now unused match information in the generated code too)
Removed type from deeplink entry and replace with whether the method param is null or not.
Needed to fix some tests as some of the tests were relying on the fact tha the Deepl;inkEntry Object in the repository and the one returned as a match are the same object. This is no longer the case as the match Entry objects are not generated on the fly.

Updated all dependencies to latest. With exception of kotlin and okio as they need some more fixes to be updated.

* Do not code gen DeepLinkEntries anymore into the registries.
Remove match index from code as it is not used anymore for matching.

Replace jcenter with mavenCentral

* Update code gen tests

* Cleanup style and tests

* Cleanup srcRoot mess.
CR changes.

* Checkstyle cleanup

* Checkstyle cleanup

Co-authored-by: Andreas Rossbacher <andreas.rossbacher@airbnb.com>